💫CHANGED💫 Updated all of the field calls on main

+++ src/utils.rs
@@ -2,7 +2,7 @@ use crate::{Dimensions, FontConfig};
 pub fn get_canvas_size(
     pref_dimensions: Option<Dimensions>,
     largest_line_length: u32,
-    num_lines: usize,
+    num_lines: u32,
     font: &FontConfig,
 ) -> (Dimensions, f32) {
     // Were using A as a reference width char as it's a good average.
